Fixed the namespace and folder of `PartialViewResultTest `; and some minor fixes.
