+++ test/Microsoft.AspNetCore.Mvc.ViewFeatures.Test/ViewResultTest.cs
@@ -55,28 +55,29 @@ namespace Microsoft.AspNetCore.Mvc
         public async Task ExecuteResultAsync_Throws_IfViewCouldNotBeFound_MessageUsesGetViewLocations()
         {
             // Arrange
+            var viewName = "MyView";
+            var actionContext = GetActionContext();
             var expected = string.Join(
                 Environment.NewLine,
-                "The view 'MyView' was not found. The following locations were searched:",
+                $"The view '{viewName}' was not found. The following locations were searched:",
                 "Location1",
                 "Location2");
 
-            var actionContext = GetActionContext();
-
             var viewEngine = new Mock<IViewEngine>(MockBehavior.Strict);
             viewEngine
-                .Setup(e => e.GetView(/*executingFilePath*/ null, "MyView", /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("MyView", new[] { "Location1", "Location2" }))
+                .Setup(v => v.GetView(/*executingFilePath*/ null, viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, Enumerable.Empty<string>()))
                 .Verifiable();
+
             viewEngine
-                .Setup(v => v.FindView(It.IsAny<ActionContext>(), It.IsAny<string>(), /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("MyView", Enumerable.Empty<string>()))
+                .Setup(v => v.FindView(It.IsAny<ActionContext>(), viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, new[] { "Location1", "Location2" }))
                 .Verifiable();
 
             var viewResult = new ViewResult
             {
                 ViewEngine = viewEngine.Object,
-                ViewName = "MyView",
+                ViewName = viewName,
                 ViewData = new ViewDataDictionary(new EmptyModelMetadataProvider()),
                 TempData = Mock.Of<ITempDataDictionary>(),
             };
@@ -92,28 +93,31 @@ namespace Microsoft.AspNetCore.Mvc
         public async Task ExecuteResultAsync_Throws_IfViewCouldNotBeFound_MessageUsesFindViewLocations()
         {
             // Arrange
+            var viewName = "MyView";
+            var actionContext = GetActionContext();
             var expected = string.Join(
                 Environment.NewLine,
-                "The view 'MyView' was not found. The following locations were searched:",
+                $"The view '{viewName}' was not found. The following locations were searched:",
                 "Location1",
-                "Location2");
-
-            var actionContext = GetActionContext();
+                "Location2",
+                "Location3",
+                "Location4");
 
             var viewEngine = new Mock<IViewEngine>(MockBehavior.Strict);
             viewEngine
-                .Setup(e => e.GetView(/*executingFilePath*/ null, "MyView", /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("MyView", Enumerable.Empty<string>()))
+                .Setup(v => v.GetView(/*executingFilePath*/ null, viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, new[] { "Location1", "Location2" }))
                 .Verifiable();
+
             viewEngine
-                .Setup(v => v.FindView(It.IsAny<ActionContext>(), It.IsAny<string>(), /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("MyView", new[] { "Location1", "Location2" }))
+                .Setup(v => v.FindView(It.IsAny<ActionContext>(), viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, new[] { "Location3", "Location4" }))
                 .Verifiable();
 
             var viewResult = new ViewResult
             {
                 ViewEngine = viewEngine.Object,
-                ViewName = "MyView",
+                ViewName = viewName,
                 ViewData = new ViewDataDictionary(new EmptyModelMetadataProvider()),
                 TempData = Mock.Of<ITempDataDictionary>(),
             };
@@ -129,30 +133,31 @@ namespace Microsoft.AspNetCore.Mvc
         public async Task ExecuteResultAsync_Throws_IfViewCouldNotBeFound_MessageUsesAllLocations()
         {
             // Arrange
+            var viewName = "MyView";
+            var actionContext = GetActionContext();
             var expected = string.Join(
                 Environment.NewLine,
-                "The view 'MyView' was not found. The following locations were searched:",
+                $"The view '{viewName}' was not found. The following locations were searched:",
                 "Location1",
                 "Location2",
                 "Location3",
                 "Location4");
 
-            var actionContext = GetActionContext();
-
             var viewEngine = new Mock<IViewEngine>(MockBehavior.Strict);
             viewEngine
-                .Setup(e => e.GetView(/*executingFilePath*/ null, "MyView", /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("MyView", new[] { "Location1", "Location2" }))
+                .Setup(v => v.GetView(/*executingFilePath*/ null, viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, new[] { "Location1", "Location2" }))
                 .Verifiable();
+
             viewEngine
-                .Setup(v => v.FindView(It.IsAny<ActionContext>(), It.IsAny<string>(), /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("MyView", new[] { "Location3", "Location4" }))
+                .Setup(v => v.FindView(It.IsAny<ActionContext>(), viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, new[] { "Location3", "Location4" }))
                 .Verifiable();
 
             var viewResult = new ViewResult
             {
                 ViewEngine = viewEngine.Object,
-                ViewName = "MyView",
+                ViewName = viewName,
                 ViewData = new ViewDataDictionary(new EmptyModelMetadataProvider()),
                 TempData = Mock.Of<ITempDataDictionary>(),
             };
@@ -168,8 +173,8 @@ namespace Microsoft.AspNetCore.Mvc
         public async Task ExecuteResultAsync_FindsAndExecutesView()
         {
             // Arrange
-            var viewName = "myview";
-            var context = GetActionContext();
+            var viewName = "MyView";
+            var actionContext = GetActionContext();
 
             var view = new Mock<IView>(MockBehavior.Strict);
             view
@@ -182,18 +187,20 @@ namespace Microsoft.AspNetCore.Mvc
                 .Setup(v => v.Dispose())
                 .Verifiable();
 
+            // Used by logging
             view
-                .Setup(v => v.Path)
-                .Returns("//location");
+                .SetupGet(v => v.Path)
+                .Returns($"{viewName}.cshtml");
 
             var viewEngine = new Mock<IViewEngine>(MockBehavior.Strict);
             viewEngine
-                .Setup(e => e.GetView(/*executingFilePath*/ null, "myview", /*isMainPage*/ true))
-                .Returns(ViewEngineResult.NotFound("myview", Enumerable.Empty<string>()))
+                .Setup(v => v.GetView(/*executingFilePath*/ null, viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.NotFound(viewName, Enumerable.Empty<string>()))
                 .Verifiable();
+
             viewEngine
-                .Setup(e => e.FindView(context, "myview", /*isMainPage*/ true))
-                .Returns(ViewEngineResult.Found("myview", view.Object))
+                .Setup(v => v.FindView(It.IsAny<ActionContext>(), viewName, /*isMainPage*/ true))
+                .Returns(ViewEngineResult.Found(viewName, view.Object))
                 .Verifiable();
 
             var viewResult = new ViewResult
@@ -205,7 +212,7 @@ namespace Microsoft.AspNetCore.Mvc
             };
 
             // Act
-            await viewResult.ExecuteResultAsync(context);
+            await viewResult.ExecuteResultAsync(actionContext);
 
             // Assert
             view.Verify();
